microservices_data preprocessing/requirements.txt
google-cloud-bigquery
pyspark
pandas
pymongo
